display: carve webrtc pool_glue.rs (pure move)
